Add files via upload
a)	Build and train a KNN Regression model. Vary the parameter K and analyze the results by identifying cases of overfitting and underfitting. Select the optimal value of K and justify your choice.
b)	Build and train an OLS Regression model. Analyze the results and indicate if the learned model is a good choice for this data. Justify your conclusions.
c)	Build and train a Ridge Regression model. Vary the constraint parameter and analyze the results by identifying cases of overfitting and underfitting. Select the optimal value of  and justify your choice.
d)	Build and train a LASSO Regression model. Vary the constraint parameter and analyze the results by identifying cases of overfitting and underfitting. Select the optimal value of  and justify your choice.